refactor

+++ src/db/schemas.py
@@ -1,9 +1,8 @@
 from pydantic import BaseModel
-from pgvector.sqlalchemy import Vector
 
 class EmbeddingSchema(BaseModel):
     name: str
-    embeddings: list[float]
+    embeddings: list[int]
 
     class Config:
         orm_mode = True
